+++ src/main/java/com/filmdoms/community/board/post/controller/PostCommentController.java
@@ -1,7 +1,7 @@
-package com.filmdoms.community.postComment.controller;
+package com.filmdoms.community.board.post.controller;
 
 import com.filmdoms.community.account.data.dto.response.Response;
-import com.filmdoms.community.postComment.service.PostCommentService;
+import com.filmdoms.community.board.post.service.PostCommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -9,14 +9,14 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/v1/post/comment")
 @RequiredArgsConstructor
 public class PostCommentController {
 
     private final PostCommentService postCommentService;
 
     // TODO: 댓글 작성 기능 구현 후 삭제
-    @PostMapping("/test/comment")
+    @PostMapping("/test-data")
     public Response testComment(@RequestParam Long postId) {
         postCommentService.testComment(postId);
         return Response.success();
