+++ src/main/java/com/filmdoms/community/board/post/controller/PostController.java
@@ -1,8 +1,8 @@
-package com.filmdoms.community.post.controller;
+package com.filmdoms.community.board.post.controller;
 
 import com.filmdoms.community.account.data.dto.response.Response;
-import com.filmdoms.community.post.data.dto.response.PostMainPageResponseDto;
-import com.filmdoms.community.post.service.PostService;
+import com.filmdoms.community.board.post.data.dto.response.PostMainPageResponseDto;
+import com.filmdoms.community.board.post.service.PostService;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -13,13 +13,13 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/v1")
+@RequestMapping("/api/v1/post")
 @RequiredArgsConstructor
 public class PostController {
 
     private final PostService postService;
 
-    @GetMapping("/main/post")
+    @GetMapping("/main-page")
     public Response<List<PostMainPageResponseDto>> viewMain() {
         return Response.success(postService.getMainPagePosts()
                 .stream()
@@ -28,7 +28,7 @@ public class PostController {
     }
 
     // TODO: 게시글 작성 기능 구현 후 삭제
-    @PostMapping("/test/post")
+    @PostMapping("/test-data")
     public Response testPost(@RequestParam String title) {
         postService.testPost(title);
         return Response.success();
