+++ src/main/java/com/filmdoms/community/board/post/data/dto/response/PostMainPageResponseDto.java
@@ -1,6 +1,6 @@
-package com.filmdoms.community.post.data.dto.response;
+package com.filmdoms.community.board.post.data.dto.response;
 
-import com.filmdoms.community.post.data.dto.PostBriefDto;
+import com.filmdoms.community.board.post.data.dto.PostBriefDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
