+++ src/main/java/com/filmdoms/community/board/post/repository/PostCommentRepository.java
@@ -1,6 +1,6 @@
-package com.filmdoms.community.postComment.repository;
+package com.filmdoms.community.board.post.repository;
 
-import com.filmdoms.community.postComment.data.entity.PostComment;
+import com.filmdoms.community.board.post.data.entity.PostComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostCommentRepository extends JpaRepository<PostComment, Long> {
