+++ src/main/java/com/filmdoms/community/board/post/repository/PostRepository.java
@@ -1,11 +1,11 @@
-package com.filmdoms.community.post.repository;
+package com.filmdoms.community.board.post.repository;
 
-import com.filmdoms.community.post.data.entity.Post;
+import com.filmdoms.community.board.post.data.entity.Post;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
 
     // 메인 페이지 호출시 사용
-    List<Post> findFirst4ByOrderByDateCreated();
+    List<Post> findFirst4ByOrderByIdDesc();
 }
